refactor(agent-scaffolders): centralize scaffold.py to plugin root and convert skills to use upward symlinks
