+++ plugins/agent-scaffolders/skills/create-docker-skill/SKILL.md
@@ -21,7 +21,7 @@ Ask the user what specific external container or pipeline orchestrator is being 
 ### 2. Scaffold the Infrastructure
 Execute the deterministic `scaffold.py` script to generate the compliant physical directories:
 ```bash
-python3 ~~agent-scaffolders-root/skills/create-docker-skill/scripts/scaffold.py --type skill --name <requested-name> --path <destination-directory> --desc "<short-description>"
+python3 ./scripts/scaffold.py --type skill --name <requested-name> --path <destination-directory> --desc "<short-description>"
 ```
 
 ### 3. Generate Pre-Flight Checker Script
