+++ plugins/agent-scaffolders/skills/create-docker-skill/scripts/scaffold.py
@@ -1,1 +1,1 @@
-../../create-skill/scripts/scaffold.py
+../../../scripts/scaffold.py
