+++ plugins/agent-scaffolders/skills/create-github-action/SKILL.md
@@ -48,7 +48,7 @@ Ask the user for the following context:
 Run the scaffold script:
 
 ```bash
-python ~~agent-scaffolders-root/scripts/scaffold_github_action.py \
+python ./scripts/scaffold_github_action.py \
   --skill-dir <path-to-skill-directory> \
   --category <test|build|lint|deploy|release|security|maintenance|custom> \
   --platform <python|nodejs|go|docker|dotnet|generic> \
