+++ plugins/agent-scaffolders/skills/create-sub-agent/scripts/scaffold.py
@@ -1,1 +1,1 @@
-../../create-skill/scripts/scaffold.py
+../../../scripts/scaffold.py
